+++ CHANGES.md
@@ -3,6 +3,7 @@
   Both show the edit dialog if the record is found, see the doc
 * Added field options listEscapeHTML and inputEscapeHTML, indicating wether or not you want the list or input value escaped
   Default is false, the old behavior
+* toolbarsearch now also supports checkbox
 
 = 3.0.11 (2025/07/01) =
 * Added logLevel option